Initial Commit
This is a poorly written, but mostly working
implementation I wrote late one night, as
one of my coworkers had taken to leaving sticky
notes on each pot of coffee to denote the
freshness, as another coworker had once complained
about a cup of cold coffee.

Theory of Operation

Do a WiFi Scan. For Each OpenAP or AP with
hardcoded credentials, connect to the AP,
and send a single NTP packet to NIST. If
NIST responds in a reasonable amount of time,
render the display, then set the ESP8266 in
deep sleep mode with no wakeup planned to
save power.

If NTP cannot be reached, shut down to save
power. This is not terribly reliable, as the
loss of a single NTP packet could cause a
failure. Original plan was to use a memory
in pixel reflective LCD and a nine-volt
battery, but the LCD was too expensive, so
with the move to OLED, I needed continuous
power anyways.